Исправил. Перевел все 4хх тесты в отдельный класс + добавил обнуление courierid.
